overflow fix

+++ strelyaev.roman/S1/inputList.cpp
@@ -18,7 +18,7 @@ void strelyaev::makeSequence(std::istream& in, List < std::pair < std::string, L
     List < size_t > list_of_nums;
     while (in >> line && !std::isalpha(line[0]))
     {
-      list_of_nums.push_back(std::stoi(line));
+      list_of_nums.push_back(std::stoull(line));
     }
     list_of_lists.push_back({name, list_of_nums});
   }
